Reverted level creation settings being tested to default values for merge into main.
