+++ Assets/ML-Agents/Timers/SampleScene_timers.json
@@ -1,1 +1,1 @@
-{"count":1,"self":7.0845823999999995,"total":7.095638,"children":{"InitializeActuators":{"count":1,"self":0.000803,"total":0.000803,"children":null},"InitializeSensors":{"count":1,"self":0.00083999999999999993,"total":0.00083999999999999993,"children":null},"AgentSendState":{"count":336,"self":0.0024059999999999997,"total":0.004968,"children":{"CollectObservations":{"count":68,"self":0.000632,"total":0.000632,"children":null},"WriteActionMask":{"count":68,"self":0.000268,"total":0.000268,"children":null},"RequestDecision":{"count":68,"self":0.0016619999999999998,"total":0.0016619999999999998,"children":null}}},"DecideAction":{"count":336,"self":0.002121,"total":0.002121,"children":null},"AgentAct":{"count":336,"self":0.0016769999999999999,"total":0.0016769999999999999,"children":null}},"gauges":{},"metadata":{"timer_format_version":"0.1.0","start_time_seconds":"1720803278","unity_version":"2022.3.35f1","command_line_arguments":"\/Applications\/Unity\/Hub\/Editor\/2022.3.35f1\/Unity.app\/Contents\/MacOS\/Unity -projectpath \/Users\/jstrong\/Programming\/CS467\/ML-Breakout -useHub -hubIPC -cloudEnvironment production -licensingIpc LicenseClient-jstrong -hubSessionId cac5ed54-7dc4-48c0-9528-98e0daed6078 -accessToken 5daoWE3THcPsQMdgscy9HuZAyIPFqWFNRp5yPF6R3Hc012f","communication_protocol_version":"1.5.0","com.unity.ml-agents_version":"2.0.1","scene_name":"SampleScene","end_time_seconds":"1720803285"}}
+{"count":1,"self":12.648689599999999,"total":12.661226,"children":{"InitializeActuators":{"count":1,"self":0.000741,"total":0.000741,"children":null},"InitializeSensors":{"count":1,"self":0.00091199999999999994,"total":0.00091199999999999994,"children":null},"AgentSendState":{"count":592,"self":0.0032329999999999998,"total":0.005604,"children":{"CollectObservations":{"count":119,"self":0.00067499999999999993,"total":0.00067499999999999993,"children":null},"WriteActionMask":{"count":119,"self":0.000278,"total":0.000278,"children":null},"RequestDecision":{"count":119,"self":0.001418,"total":0.001418,"children":null}}},"DecideAction":{"count":592,"self":0.0024649999999999997,"total":0.0024649999999999997,"children":null},"AgentAct":{"count":592,"self":0.002157,"total":0.002157,"children":null}},"gauges":{},"metadata":{"timer_format_version":"0.1.0","start_time_seconds":"1721060440","unity_version":"2022.3.35f1","command_line_arguments":"\/Applications\/Unity\/Hub\/Editor\/2022.3.35f1\/Unity.app\/Contents\/MacOS\/Unity -projectpath \/Users\/jstrong\/Programming\/CS467\/ML-Breakout -useHub -hubIPC -cloudEnvironment production -licensingIpc LicenseClient-jstrong -hubSessionId cac5ed54-7dc4-48c0-9528-98e0daed6078 -accessToken C7Pp9WmAYMOtyjGbSa-cevIcx94LlnMY8RdA9rHKZGA012f","communication_protocol_version":"1.5.0","com.unity.ml-agents_version":"2.0.1","scene_name":"SampleScene","end_time_seconds":"1721060452"}}
