+++ Assets/Scenes/MainMenu.unity
@@ -383,7 +383,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 137572539}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 81817147a22a047b2bb852ffe1cac97f, type: 3}
+  m_Script: {fileID: 11500000, guid: c974bd2dfb1774cbe889e85237de6fb4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!114 &137572542
@@ -430,7 +430,7 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 137572539}
+      - m_Target: {fileID: 53665331}
         m_TargetAssemblyTypeName: 
         m_MethodName: 
         m_Mode: 1
@@ -2292,50 +2292,6 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1290547846}
   m_CullTransparentMesh: 1
---- !u!1 &1401609651
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 1401609653}
-  - component: {fileID: 1401609652}
-  m_Layer: 0
-  m_Name: CoverImageManager
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!114 &1401609652
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1401609651}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 81817147a22a047b2bb852ffe1cac97f, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
---- !u!4 &1401609653
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1401609651}
-  serializedVersion: 2
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -2.327691, y: -4.720888, z: 11.861219}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_ConstrainProportionsScale: 0
-  m_Children: []
-  m_Father: {fileID: 0}
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1660057539 &9223372036854775807
 SceneRoots:
   m_ObjectHideFlags: 0
@@ -2344,4 +2300,3 @@ SceneRoots:
   - {fileID: 53665336}
   - {fileID: 809236950}
   - {fileID: 1246044836}
-  - {fileID: 1401609653}
