--- Assets/Scripts/CoverImageHandler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 81817147a22a047b2bb852ffe1cac97f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
